refactors article endpoint into namespaced api
