+++ Daydream/Networking/API.swift
@@ -12,3 +12,5 @@ import CoreLocation
 
 // Creates namespace for API calls. Routes for each namespaced API are in extensions.
 enum API {}
+
+typealias SightsAndEateries = ([Place], [Eatery])
